+++ sorting/merge_sort/src/merge_sort.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 #include<stdlib.h>
 
-void merge(int* nums, int l, int m, int r, int* tmp) {
-    int i1 = 0, i2 = 0, i = 0;
-    int n1 = m-l+1, n2 = r-m;
+void merge(int* nums, size_t l, size_t m, size_t r, int* tmp) {
+    size_t i1 = 0, i2 = 0, i = 0;
+    size_t n1 = m-l+1, n2 = r-m;
 
     while(i1<n1 || i2<n2) {
         if(i1==n1 || (i2<n2 && nums[m+1+i2]<nums[l+i1])) {
@@ -16,7 +16,7 @@ void merge(int* nums, int l, int m, int r, int* tmp) {
     for(size_t i=l;i<=r;i++) nums[i]=tmp[i];
 }
 
-void mergeSort(int* nums, int l, int r, int* tmp) {
+void mergeSort(int* nums, size_t l, size_t r, int* tmp) {
     if(l>=r) return;
 
     int m=l+(r-l)/2;
